moved dynamics code into its own subdirectory. Added flag to graspit.pro to allow graspit to build either with graspit_dynamics or bullet_dynamics. Both compile and run fine.
