+++ include/dynamics/bulletDynamics.h
@@ -17,7 +17,7 @@
 // You should have received a copy of the GNU General Public License
 // along with GraspIt!.  If not, see <http://www.gnu.org/licenses/>.
 //
-// Author(s): Matei T. Ciocarlie
+// Author(s): Matei T. Ciocarlie, Jake Varley
 //
 // $Id:$
 //
